Addressing comments left on PR. Taken interface and other classes other of DbUtils.kt and put them in their own separate files.
